[Malterlib/Core] Disable thread safe statics on Windows
Thread safe statics support in MSVC Crt is uninitialized as part of C exit. This can cause exceptions when shutting down Malterlib because statics are initialized as part of this. This in turn is caught by the thread cleanup in Windows causing memory leaks or missed destructors causing deadlocks.

Copied from Perforce
 Change: 92794